// frontend/src/app/models/product.ts
export interface Product {
    _id?:string;
    productName: string;
    productShortCode: string;
    category: string;
    price: number;
    quantity: number;
    description: string,
    imageUrl: string,
    isBestAchieved: boolean;
    createdDate: Date;
    origin: string;
}

export interface Cart {
    _id?:string;
    productId: Product;
    quantity:number;
}